bespoke insert implementation
