+++ crates/datafusion_ext/src/lib.rs
@@ -3,7 +3,6 @@ pub mod metrics;
 pub mod planner;
 pub mod runtime;
 pub mod session_metrics;
-pub mod stream;
 pub mod vars;
 pub use planner::*;
 pub mod functions;
